+++ actix-web-lab-derive/tests/trybuild/err-invalid-structures.rs
@@ -1,0 +1,11 @@
+use actix_web_lab::FromRequest;
+
+#[derive(FromRequest)]
+enum Foo {
+    Data(()),
+}
+
+#[derive(FromRequest)]
+struct Bar(());
+
+fn main() {}