+++ actix-web-lab-derive/tests/trybuild/ok-with-body-type.rs
@@ -1,0 +1,12 @@
+use std::collections::HashMap;
+
+use actix_web::{http, web};
+use actix_web_lab_derive::FromRequest;
+
+#[derive(Debug, FromRequest)]
+struct RequestParts {
+    pool: web::Data<u32>,
+    form: web::Json<HashMap<String, String>>,
+}
+
+fn main() {}